+++ example/lib/samples/context_extension_view.dart
@@ -63,7 +63,11 @@ class ContextExtensionView extends StatelessWidget {
   }
 
   Widget radiusExtension(BuildContext context) {
+<<<<<<< HEAD
     return Container(
+=======
+    return DecoratedBox(
+>>>>>>> master
       decoration: BoxDecoration(borderRadius: context.lowBorderRadius),
     );
   }