+++ example/lib/samples/model/post.dart
@@ -1,4 +1,8 @@
+<<<<<<< HEAD
 import 'package:kartal/src/utility/bundle/INetworkModel.dart';
+=======
+import 'package:kartal/src/utility/bundle/i_network_model.dart';
+>>>>>>> master
 
 class Post extends INetworkModel<Post> {
   Post({
@@ -7,6 +11,10 @@ class Post extends INetworkModel<Post> {
     this.title,
     this.body,
   });
+<<<<<<< HEAD
+=======
+
+>>>>>>> master
   factory Post.fromJson(Map<String, dynamic> json) {
     return Post(
       userId: json['userId'] as int?,
@@ -15,6 +23,10 @@ class Post extends INetworkModel<Post> {
       body: json['body'] as String?,
     );
   }
+<<<<<<< HEAD
+=======
+
+>>>>>>> master
   int? userId;
   int? id;
   String? title;