+++ lib/kartal.dart
@@ -4,13 +4,13 @@ library kartal;
 
 export 'package:collection/collection.dart';
 
-export './src/constants/index.dart';
-export './src/index.dart';
+export 'src/constants/index.dart';
 export 'src/constants/input_formatter_constants.dart';
 export 'src/context_extension.dart';
 export 'src/file_extension.dart';
 export 'src/future_extension.dart';
 export 'src/image_extension.dart';
+export 'src/index.dart';
 export 'src/int_extension.dart';
 export 'src/list_extension.dart';
 export 'src/string_extension.dart';