+++ lib/src/constants/index.dart
@@ -2,4 +2,4 @@ export 'app_constants.dart';
 export 'file_constants.dart';
 export 'input_formatter_constants.dart';
 export 'regex_constants.dart';
-export 'responsivity_constants.dart';
+export 'responsibility_constants.dart';