+++ lib/src/constants/regex_constants.dart
@@ -5,6 +5,8 @@ class RegexConstants {
   RegexConstants._init();
   static RegexConstants? _instance;
 
-  final String emailRegex = r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$';
-  final String passwordRegex = r'^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[!@#\$&*~]).{8,}$';
+  final String emailRegex =
+      r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9\w-]+\.[a-zA-Z]+";
+  final String passwordRegex =
+      r'^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[!@#\$&*~]).{8,}$';
 }