+++ lib/src/list_extension.dart
@@ -8,4 +8,11 @@ extension ListExtension<T> on List<T>? {
       return false;
     }
   }
+
+  /// It return index when find a item instead of -1
+  /// -1 means not found and it comes from core library
+  int? indexOrNull(bool Function(T) search) {
+    final result = this?.indexWhere(search);
+    return result != -1 ? result : null;
+  }
 }