+++ lib/src/utility/bundle/bundle_decoder.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:kartal/src/utility/bundle/INetworkModel.dart';
+import 'package:kartal/src/utility/bundle/i_network_model.dart';
 
 typedef ComputeCallback<T extends INetworkModel<T>, R> = Future<R> Function(
   // ignore: library_private_types_in_public_api