+++ lib/src/utility/device_utility.dart
@@ -1,12 +1,19 @@
 // ignore_for_file: prefer_constructors_over_static_methods
 
-import 'package:device_info/device_info.dart';
+import 'dart:io';
+
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:kartal/src/constants/app_constants.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class DeviceUtility {
   DeviceUtility._init();
+
+  static Future<void> deviceInit() async {
+    await instance.initPackageInfo();
+  }
+
   static DeviceUtility? _instance;
   static DeviceUtility get instance {
     if (_instance != null) {
@@ -24,7 +31,9 @@ class DeviceUtility {
 
   Future<bool> isIpad() async {
     info = await deviceInfo.iosInfo;
-    return info.name.toLowerCase().contains(KartalAppConstants.IPAD_TYPE);
+    return (info.name ?? '')
+        .toLowerCase()
+        .contains(KartalAppConstants.IPAD_TYPE);
   }
 
   String shareMailText(String title, String body) {
@@ -33,5 +42,18 @@ class DeviceUtility {
 
   Future<void> initPackageInfo() async {
     packageInfo = await PackageInfo.fromPlatform();
+  }
+
+  Future<String> getUniqueDeviceId() async {
+    if (Platform.isIOS) {
+      final iosDeviceInfo = await deviceInfo.iosInfo;
+      return iosDeviceInfo.identifierForVendor ?? '';
+    } else if (Platform.isAndroid) {
+      final androidDeviceInfo = await deviceInfo.androidInfo;
+
+      return androidDeviceInfo.id;
+    } else {
+      throw Exception('Thats not supported other platform');
+    }
   }
 }